Fix some selection intent not overridable in EditableText.
Before the change, the action handles DirectionalCaretMovementIntent, if
user override ExtendSelectionVerticallyToAdjacentLineIntent or
ExtendSelectionVerticallyToAdjacentPageIntent, it doesn't work because
overridable actions are checked against the concrete type.

The ideal fix is to change `_castAction` in `actions.dart` to allow
actions that can parent type too. However, I haven't figure out how to
do it, so this fix is made to mitigate the issue. This fixes allows Line
and Page to be two actions, but share the same run.